requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

aiohttp==3.8.3
    # via
    #   eth-brownie
    #   web3
aiosignal==1.2.0
    # via
    #   aiohttp
    #   eth-brownie
argon2-cffi==21.3.0
    # via passlib
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via
    #   eth-brownie
    #   vyper
async-timeout==4.0.2
    # via
    #   aiohttp
    #   eth-brownie
attrs==22.1.0
    # via
    #   aiohttp
    #   eth-brownie
    #   hypothesis
    #   jsonschema
    #   pytest
base58==2.1.1
    # via
    #   eth-brownie
    #   multiaddr
bitarray==2.6.0
    # via
    #   eth-account
    #   eth-brownie
black==22.10.0
    # via eth-brownie
certifi==2022.9.24
    # via
    #   eth-brownie
    #   requests
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   eth-brownie
    #   requests
click==8.1.3
    # via
    #   black
    #   eth-brownie
    #   flask
    #   uvicorn
cryptography==39.0.1
    # via pyseto
cytoolz==0.12.0
    # via
    #   eth-brownie
    #   eth-keyfile
    #   eth-utils
dataclassy==0.11.1
    # via
    #   eip712
    #   eth-brownie
eip712==0.1.0
    # via eth-brownie
eth-abi==2.2.0
    # via
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   web3
eth-account==0.5.9
    # via
    #   eth-brownie
    #   web3
eth-brownie==1.19.3
    # via -r requirements.in
eth-event==1.2.3
    # via eth-brownie
eth-hash[pycryptodome]==0.3.3
    # via
    #   eth-brownie
    #   eth-event
    #   eth-utils
    #   web3
eth-keyfile==0.5.1
    # via
    #   eth-account
    #   eth-brownie
eth-keys==0.3.4
    # via
    #   eth-account
    #   eth-brownie
    #   eth-keyfile
eth-rlp==0.2.1
    # via
    #   eth-account
    #   eth-brownie
    #   web3
eth-typing==2.3.0
    # via
    #   eip712
    #   eth-abi
    #   eth-brownie
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==1.10.0
    # via
    #   eip712
    #   eth-abi
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
execnet==1.9.0
    # via
    #   eth-brownie
    #   pytest-xdist
flask==2.2.3
    # via -r requirements.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
    #   eth-brownie
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
hexbytes==0.2.3
    # via
    #   -r requirements.in
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-rlp
    #   web3
hypothesis==6.27.3
    # via eth-brownie
idna==3.4
    # via
    #   eth-brownie
    #   requests
    #   yarl
importlib-metadata==6.0.0
    # via flask
inflection==0.5.0
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
iniconfig==1.1.1
    # via
    #   eth-brownie
    #   pytest
ipfshttpclient==0.8.0a2
    # via
    #   eth-brownie
    #   web3
iso8601==1.1.0
    # via pyseto
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jsonschema==3.2.0
    # via
    #   eth-brownie
    #   mythx-models
    #   web3
lazy-object-proxy==1.7.1
    # via eth-brownie
lru-dict==1.1.8
    # via
    #   eth-brownie
    #   web3
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
multiaddr==0.0.9
    # via
    #   eth-brownie
    #   ipfshttpclient
multidict==6.0.2
    # via
    #   aiohttp
    #   eth-brownie
    #   yarl
mypy-extensions==0.4.3
    # via
    #   black
    #   eth-brownie
mythx-models==1.9.1
    # via
    #   eth-brownie
    #   pythx
netaddr==0.8.0
    # via
    #   eth-brownie
    #   multiaddr
packaging==21.3
    # via
    #   eth-brownie
    #   pytest
parsimonious==0.8.1
    # via
    #   eth-abi
    #   eth-brownie
passlib[argon2]==1.7.4
    # via pyseto
pathspec==0.10.1
    # via
    #   black
    #   eth-brownie
platformdirs==2.5.2
    # via
    #   black
    #   eth-brownie
pluggy==1.0.0
    # via
    #   eth-brownie
    #   pytest
prompt-toolkit==3.0.31
    # via eth-brownie
protobuf==3.19.5
    # via
    #   eth-brownie
    #   web3
psutil==5.9.2
    # via eth-brownie
py==1.11.0
    # via
    #   eth-brownie
    #   pytest
    #   pytest-forked
py-solc-ast==1.2.9
    # via eth-brownie
py-solc-x==1.1.1
    # via eth-brownie
pycparser==2.21
    # via cffi
pycryptodome==3.15.0
    # via
    #   eip712
    #   eth-brownie
    #   eth-hash
    #   eth-keyfile
    #   vyper
pycryptodomex==3.17
    # via pyseto
pygments==2.13.0
    # via
    #   eth-brownie
    #   pygments-lexer-solidity
pygments-lexer-solidity==0.7.0
    # via eth-brownie
pyjwt==1.7.1
    # via
    #   eth-brownie
    #   pythx
pyparsing==3.0.9
    # via
    #   eth-brownie
    #   packaging
pyrsistent==0.18.1
    # via
    #   eth-brownie
    #   jsonschema
pyseto==1.7.1
    # via -r requirements.in
pytest==6.2.5
    # via
    #   eth-brownie
    #   pytest-forked
    #   pytest-xdist
pytest-forked==1.4.0
    # via
    #   eth-brownie
    #   pytest-xdist
pytest-xdist==1.34.0
    # via eth-brownie
python-dateutil==2.8.1
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
python-dotenv==0.16.0
    # via eth-brownie
pythx==1.6.1
    # via eth-brownie
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   eth-brownie
requests==2.28.1
    # via
    #   eth-brownie
    #   ipfshttpclient
    #   py-solc-x
    #   pythx
    #   vvm
    #   web3
rlp==2.0.1
    # via
    #   -r requirements.in
    #   eth-account
    #   eth-brownie
    #   eth-rlp
semantic-version==2.10.0
    # via
    #   eth-brownie
    #   py-solc-x
    #   vvm
    #   vyper
six==1.16.0
    # via
    #   asttokens
    #   eth-brownie
    #   jsonschema
    #   multiaddr
    #   parsimonious
    #   pytest-xdist
    #   python-dateutil
sortedcontainers==2.4.0
    # via
    #   eth-brownie
    #   hypothesis
toml==0.10.2
    # via
    #   eth-brownie
    #   pytest
tomli==2.0.1
    # via
    #   black
    #   eth-brownie
toolz==0.12.0
    # via
    #   cytoolz
    #   eth-brownie
tqdm==4.64.1
    # via eth-brownie
types-pyyaml==6.0.12.8
    # via -r requirements.in
typing-extensions==4.4.0
    # via
    #   black
    #   eth-brownie
urllib3==1.26.12
    # via
    #   eth-brownie
    #   requests
uvicorn==0.20.0
    # via -r requirements.in
varint==1.0.2
    # via
    #   eth-brownie
    #   multiaddr
vvm==0.1.0
    # via eth-brownie
vyper==0.3.7
    # via eth-brownie
wcwidth==0.2.5
    # via
    #   eth-brownie
    #   prompt-toolkit
web3==5.31.3
    # via
    #   -r requirements.in
    #   eth-brownie
websockets==9.1
    # via
    #   eth-brownie
    #   web3
werkzeug==2.2.3
    # via flask
wheel==0.37.1
    # via
    #   eth-brownie
    #   vyper
wrapt==1.14.1
    # via eth-brownie
yarl==1.8.1
    # via
    #   aiohttp
    #   eth-brownie
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
